Add Matrix `.well-know` JSON files
These files are required to host a Matrix server under the
`tockos.org` domain name. The server does not need to be accessible on
this domain (in this case, we chose to use `matrix.tockos.org`), but
this allows us to create room addresses like `#general:tockos.org` and
user addresses such as `@lschuermann:tockos.org`.